[FAIL] Re-enable Babbage-era minimum UTxO golden tests.
Note that these tests currently fail, so we have marked this commit with
the [FAIL] tag.

Subsequent commits will also be marked with [FAIL] until the whole test
suite passes. The first commit that causes the whole test suite to pass
will be marked with [PASS].

Test failures are of the following form:
```
  lib/shelley/test/unit/Cardano/Wallet/Shelley/MinimumUTxOSpec.hs:496:9:
  1) computeMinimumCoinForUTxO, Golden Tests, goldenTests_computeMinimumCoinForUTxO Babbage, golden test #0
       Falsified (after 1 test):
         resultExpected:
             Coin 995610

         resultReturned:
             Coin 1077500

         Condition violated: resultReturned == resultExpected

  To rerun use: --match "/computeMinimumCoinForUTxO/Golden Tests/goldenTests_computeMinimumCoinForUTxO Babbage/golden test #0/"
  ...
```